fix: go import errors due to invalid imports
After carefully looking through the package I concluded that the imports in
`tools.go` (which are causing the issues) are only used by the build system
and are not actually needed by the code in this package.

Reading [this](https://github.com/golang/go/wiki/Modules#how-can-i-track-tool-dependencies-for-a-module)
article I fully understand the reasoning behind it, but it does cause some
unneeded import errors (see screenshots below).

The article notes that this solution (adding build dependencies to a `tools.go`
file) was a best practice for Go 1.11, but it also mentions a different solution
for Go >=1.16 which doesn't has the same side effects and additionally makes the
`go.mod` file a lot smaller and cleaner.

So I updated the code to use the newer solution, making sure no stability
guarantees are lost in the process (the tools still use the exact same versions).
Doing so prevents the `go.mod` and Go import errors in projects that (need to)
import `herodot` (either directly or by importing other Ory packages that rely
on `herodot`).

It would be really awesome if this can get merged to get rid of the errors in
our build setup. Thanks!!